+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Output stream with an arbitrary number of channels (`glyphtune.output.Stream`).
 - Effects module (including base `Effect` object).
 - Stereo effects module (including `StereoPan`, `StereoLevels`, `StereoInterMix`, and `StereoDelay`).
+- Output stream with an arbitrary number of channels (`glyphtune.output.Stream`).
+- Output to wav file (`glyphtune.output.write_wav` and `glyphtune.output.WavParameters`)
 - `glyphtune.arrays` module (including `FloatArray` type alias and `is_stereo_signal` function).
 
 ### Changed